refactor: API 호출 로직을 서비스 레이어로 분리
